+++ JuiceMaker/JuiceMaker/Controller/JuiceOrderButton.swift
@@ -5,7 +5,6 @@
 //  Created by 천수현 on 2021/03/12.
 //
 
-import Foundation
 import UIKit
 class JuiceOrderButton: UIButton {
    var juice: Juice = JuiceType.strawberry
