+++ JuiceMaker/JuiceMaker/Controller/StockLabel.swift
@@ -5,7 +5,6 @@
 //  Created by 천수현 on 2021/03/11.
 //
 
-import Foundation
 import UIKit
 
 class StockLabel: UILabel {
