+++ JuiceMaker/JuiceMaker/Model/Fruit.swift
