+++ JuiceMaker/JuiceMaker/Model/Juice.swift
