+++ JuiceMaker/JuiceMaker/Model/JuiceMaker.swift
@@ -4,8 +4,6 @@
 //  Copyright © yagom academy. All rights reserved.
 // 
 
-import Foundation
-
 /// 쥬스 메이커 타입
