+++ JuiceMaker/JuiceMaker/Model/JuiceRequirement.swift
@@ -5,7 +5,6 @@
 //  Created by 천수현 on 2021/03/08.
 //
 
-
 enum Requirements {
     static let strawberry = [(FruitType.strawberry, 16)]
     static let banana = [(FruitType.banana, 2)]
