+++ JuiceMaker/JuiceMaker/View/StockView.swift
@@ -1,0 +1,35 @@
+//
+//  StockViewController.swift
+//  JuiceMaker
+//
+//  Created by 천수현 on 2021/03/14.
+//
+
+import UIKit
+
+class StockViewController: UIViewController {
+    
+    @IBOutlet weak var strawberryStock: StockLabel!
+    @IBOutlet weak var bananaStock: StockLabel!
+    @IBOutlet weak var pineappleStock: StockLabel!
+    @IBOutlet weak var kiwiStock: StockLabel!
+    @IBOutlet weak var mangoStock: StockLabel!
+    
+    @IBOutlet weak var strawBerryStockStepper: UIStepper!
+    @IBOutlet weak var bananaStockStepper: UIStepper!
+    @IBOutlet weak var pineappleStockStepper: UIStepper!
+    @IBOutlet weak var kiwiStockStepper: UIStepper!
+    @IBOutlet weak var mangoStockStepper: UIStepper!
+    
+    
+    /*
+     // MARK: - Navigation
+     
+     // In a storyboard-based application, you will often want to do a little preparation before navigation
+     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+     // Get the new view controller using segue.destination.
+     // Pass the selected object to the new view controller.
+     }
+     */
+    
+}
